Clean uo the code
