+++ tests/config.js
@@ -4,7 +4,7 @@ const domain = env.DOMAIN || 'local.kyma.dev';
 module.exports = {
   domain: domain,
   localDev: env.LOCAL_DEV || false,
-  namespace: env.NAMESPACE_NAME,
+  namespaceName: env.NAMESPACE_NAME,
   clusterAddress: env.LOCAL_DEV
     ? `http://localhost:4200/clusters`
     : `https://busola.${domain}/clusters`,
