+++ tests/support/commands.js
@@ -1,14 +1,6 @@
 import 'cypress-file-upload';
 import config from '../config';
 
-Cypress.Commands.add('handleInvalidLoginData', () => {
-  const loginErrorAlert = Cypress.$('#error');
-  if (loginErrorAlert.length !== 0) {
-    throw Error(`Login failed with message: ${loginErrorAlert.text()}`);
-  }
-  return cy.end();
-});
-
 Cypress.Commands.add(
   'shouldHaveTrimmedText',
   { prevSubject: true },
@@ -17,76 +9,14 @@ Cypress.Commands.add(
     return subject;
   },
 );
-
-const getLeftNav = () => cy.get('nav[data-testid=semiCollapsibleLeftNav]');
-Cypress.Commands.add('loginAndSelectCluster', () => {
-  console.log('env', config);
-  cy.visit(config.clusterAddress)
-    .getIframeBody()
-    .contains('Add Cluster')
-    .click();
-
-  cy.getIframeBody()
-    .contains('Drag file here')
-    .attachFile('kubeconfig.yaml', { subjectType: 'drag-n-drop' });
-
-  getLeftNav()
-    .contains('Namespaces')
-    .click();
-
-  cy.getIframeBody()
-    .find('[role=alert]')
-    .should('not.exist');
-  cy.url().should('match', /namespaces$/);
-  cy.getIframeBody()
-    .find('thead')
-    .should('be.visible'); //wait for the namespaces XHR request to finish to continue running the tests. There's no <thead> while the request is pending.
-
-  return cy.end();
-});
 
 Cypress.Commands.add('goToNamespaceDetails', () => {
   // // Go to the details of namespace
   cy.getIframeBody()
-    .contains('a', config.namespace)
+    .contains('a', config.namespaceName)
     .click();
 
   return cy.end();
-});
-
-Cypress.Commands.add(
-  'shouldHaveTrimmedText',
-  { prevSubject: true },
-  (subject, equalTo) => {
-    expect(subject.text().trim()).to.eq(equalTo);
-    return subject;
-  },
-);
-
-let LOCAL_STORAGE_MEMORY = {};
-let LAST_URL;
-
-Cypress.Commands.add('saveLocalStorageCache', () => {
-  Object.keys(localStorage).forEach(key => {
-    LOCAL_STORAGE_MEMORY[key] = localStorage[key];
-  });
-});
-
-Cypress.Commands.add('restoreLocalStorageCache', () => {
-  Object.keys(LOCAL_STORAGE_MEMORY).forEach(key => {
-    localStorage.setItem(key, LOCAL_STORAGE_MEMORY[key]);
-  });
-});
-
-Cypress.Commands.add('saveURL', () => {
-  // console.log('save url', cy.url());
-  cy.url().then(url => {
-    LAST_URL = url;
-  });
-});
-
-Cypress.Commands.add('restoreURL', () => {
-  cy.visit(LAST_URL);
 });
 
 /**
