--- tests/support/login.js
@@ -1,7 +1,0 @@
-beforeEach(() => {
-  cy.restoreLocalStorageCache();
-});
-
-afterEach(() => {
-  cy.saveLocalStorageCache();
-});
