+++ tests/tests/create-a-function.spec.js
@@ -100,10 +100,6 @@ context('Busola - Create a Function and access it', () => {
   });
 
   it('Create an API Rule for the Function', () => {
-    // getLeftNav()
-    //   .contains('Discovery and Network')
-    //   .click();
-
     getLeftNav()
       .contains('API Rules')
       .click({ force: true });
